Move Triggers from sdk-core to runners-core-construction
Converting to/from Runner API is not the SDK API surface.
